Refacto Dashboard component
