+++ src/components/Dashboard/dashboard.scss
@@ -1,4 +1,4 @@
-@import 'src/styles/vars';
+@import '@/styles/vars';
 
 #dashboardSegment{
     background: transparent;
